fix text component + anchor link
