+++ src/components/Autosuggest/index.js
@@ -1,11 +1,11 @@
 /** @jsx jsx */
-import { jsx, css } from '@emotion/core'
+import { jsx } from '@emotion/core'
 import React from 'react'
 import Autosuggest from 'react-autosuggest'
 import AutosuggestHighlightMatch from 'autosuggest-highlight/match'
 import AutosuggestHighlightParse from 'autosuggest-highlight/parse'
-import Text from 'components/Text/Text'
-import { TEXT_TYPE } from 'components/Text/Text.styles'
+import Text from 'components/Text'
+import { TEXT_TYPE } from 'components/Text/constants'
 import styles from './Autosuggest.styles'
 
 const getSuggestionValue = suggestion => suggestion.name
