+++ src/components/Text/constants.js
@@ -1,0 +1,9 @@
+export const TEXT_TYPE = {
+    'HEADER_1': 'HEADER_1',
+    'HEADER_2': 'HEADER_2',
+    'HEADER_3': 'HEADER_3',
+    'HEADER_4': 'HEADER_4',
+    'BODY_1': 'BODY_1',
+    'BODY_2': 'BODY_2',
+    'NOTE': 'NOTE',
+  }
